+++ formula-android/src/main/java/com/instacart/formula/android/events/FragmentLifecycleEvent.kt
@@ -1,4 +1,4 @@
-package com.instacart.formula.fragment
+package com.instacart.formula.android.events
 
 import com.instacart.formula.android.FragmentId