+++ formula-android/src/main/java/com/instacart/formula/android/internal/FragmentLifecycle.kt
@@ -7,7 +7,7 @@ import com.instacart.formula.android.FragmentId
 import com.instacart.formula.android.FragmentKey
 import com.instacart.formula.android.BaseFormulaFragment
 import com.instacart.formula.android.FormulaFragment
-import com.instacart.formula.fragment.FragmentLifecycleEvent
+import com.instacart.formula.android.events.FragmentLifecycleEvent
 
 /**
  * Provides utility method [lifecycleEvents] to track what fragments are added and removed from the backstack.